% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndjson.R
\name{flatten}
\alias{flatten}
\title{Flatten a character vector of individual JSON lines into a \code{data.table}}
\usage{
flatten(x, cls = c("dt", "tbl"))
}
\arguments{
\item{x}{character vector of individual JSON lines to flatten}

\item{cls}{the package uses \code{data.table::rbindlist} for speed but
that's not always the best return type for everyone, so you have
option of keeping it a \code{data.table} or converting it to a \code{tbl}}
}
\value{
\code{data.table} or \code{tbl}
}
\description{
Flatten a character vector of individual JSON lines into a \code{data.table}
}
\examples{
flatten('{"top":{"next":{"final":1,"end":true},"another":"yes"},"more":"no"}')
}
