% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndjson-package.R
\docType{package}
\name{ndjson}
\alias{ndjson}
\title{Wicked-fast Streaming JSON ('ndjson) Reader}
\description{
Streaming 'JSON' ('ndjson') has one 'JSON' record per-line
and many modern 'ndjson' files contain large numbers of records.
These constructs may not be columnar in nature, but it is often
useful to read in these files and "flatten" the structure out to
enable working with the data in an R 'data.frame'-like context.
Functions are provided that make it possible to read in plain
ndjson' files or compressed ('gz') 'ndjson' files and either
validate the format of the records or create "flat" 'data.table'
structures from them.
}
\author{
Bob Rudis (bob@rud.is)
}
