% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndjson.R
\name{stream_in}
\alias{stream_in}
\title{Stream in JSON from a character vector}
\usage{
stream_in(x)
}
\arguments{
\item{x}{character vector}
}
\description{
Given a character vector of streaming JSON (ndjson) this function will
create a flat \code{data.table} / \code{tbl_dt} from it.
}

