% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndjson.R
\name{stream_in_file}
\alias{stream_in_file}
\title{Stream in JSON from a file}
\usage{
stream_in_file(x)
}
\arguments{
\item{x}{path}
}
\description{
Given a file of streaming JSON (ndjson) this function uses \code{stringi::stri_read_lines()}
to read the data in quickly and create a flat \code{data.table} / \code{tbl_dt}
from it.
}

