% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndjson.R
\name{flatten}
\alias{flatten}
\title{Flatten a character vector of individual JSON lines into a \code{tbl_dt}}
\usage{
flatten(x)
}
\arguments{
\item{x}{character vector of individual JSON lines to flatten}
}
\value{
\code{tbl_dt}
}
\description{
Flatten a character vector of individual JSON lines into a \code{tbl_dt}
}
\examples{
flatten('{"top":{"next":{"final":1,"end":true},"another":"yes"},"more":"no"}')
}
