% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndjson-package.R
\docType{package}
\name{ndjson}
\alias{ndjson}
\alias{ndjson-package}
\title{Wicked-fast Streaming JSON ('ndjson) Reader}
\description{
Wicked-fast Streaming JSON ('ndjson) Reader
}
\author{
Bob Rudis (@hrbrmstr)
}

