% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndjson.R
\name{validate}
\alias{validate}
\title{Validate ndjson file}
\usage{
validate(path, verbose = FALSE)
}
\arguments{
\item{path}{path to file (supports "\code{gz}" files)}

\item{verbose}{display verbose information (filename and line numbers with bad records)}
}
\value{
logical
}
\description{
Given a file of streaming JSON (ndjson) this function reads in the records
and validates that they are all legal JSON records
}

